pg_control: fix reading of pg_control of vanilla in pg_controldata
Do not change the CRC as this makes its check senseless. Instead, read the
pg_control of vanilla to the pg_control strcucture of vanilla
(ControlFileDataOriginal) and print its values as is done in vanilla.

Thanks to Arthur Zakirov for reporting this.

Also: fix checking of byte ordering mismatch of pg_control in pg_probackup.