grab_excl_lock_file: don't use fio_*
This function checks for concurrent locker by "kill" command, which is
strictly local. There is no way to make it remote reliably.
More over, we have to write pid of remote agent.
So, if for whatever reason we will need to lock backup on remote host,
we'd better call this function from agent.

And, it will be better to use fcntl(F_SETLK) on Unix and
LockFileEx on Windows. But lets leave it for future.