[PBCKP-395] use pioOpenRewrite in archive.c: push_file_internal
As a unfortunate side effect it leads to absence of concurrent archiving
detection. This way "last-win" happens.

S3 gives no way to solve the issue (and this change is preparatino for S3).
But we could use file-locking if storage is on-disk. So push_file_internal
should be re-improved to use it.

test_archive_push_part_file_exists_not_stale is disabled and should be
rewritten to accompany rewritten push_file_internal with file-locking.

test_archive_push_partial_file_exists is removed completely.